Move ConfigureProgressDialog in own file
